# [Enhancement] add view def cache and remove unnecessary exception thrown