--- a/fe/fe-core/src/main/java/com/starrocks/alter/AlterJobExecutor.java
+++ b/fe/fe-core/src/main/java/com/starrocks/alter/AlterJobExecutor.java
@@ -15,6 +15,7 @@
 package com.starrocks.alter;
 
 import com.google.common.base.Preconditions;
+import com.google.common.collect.ImmutableList;
 import com.starrocks.analysis.ParseNode;
 import com.starrocks.analysis.TableName;
 import com.starrocks.catalog.Database;
@@ -28,6 +29,7 @@ import com.starrocks.qe.ConnectContext;
 import com.starrocks.scheduler.mv.MaterializedViewMgr;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.server.LocalMetastore;
+import com.starrocks.server.MetadataMgr;
 import com.starrocks.sql.analyzer.SemanticException;
 import com.starrocks.sql.ast.AddColumnClause;
 import com.starrocks.sql.ast.AddColumnsClause;
@@ -142,6 +144,9 @@ public class AlterJobExecutor extends AstVisitor<Void, ConnectContext> {
                     String.format("based table %s swapped", origTblName));
             LocalMetastore.inactiveRelatedMaterializedView(db, olapNewTbl,
                     String.format("based table %s swapped", newTblName));
+
+            MetadataMgr.clearViewAnalyzedCache(ImmutableList.of(new TableName(db.getOriginName(), newTblName),
+                    new TableName(db.getOriginName(), origTblName)));
 
             SwapTableOperationLog log = new SwapTableOperationLog(db.getId(), origTable.getId(), olapNewTbl.getId());
             GlobalStateMgr.getCurrentState().getAlterJobMgr().swapTableInternal(log);