--- a/fe/fe-core/src/main/java/com/starrocks/connector/iceberg/ScalarOperatorToIcebergExpr.java
+++ b/fe/fe-core/src/main/java/com/starrocks/connector/iceberg/ScalarOperatorToIcebergExpr.java
@@ -46,6 +46,7 @@ import java.time.ZoneId;
 import java.time.ZoneOffset;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
@@ -318,43 +319,48 @@ public class ScalarOperatorToIcebergExpr {
         }
 
         private ConstantOperator tryCastToResultType(ConstantOperator operator, Type.TypeID resultTypeID) {
-            try {
-                switch (resultTypeID) {
-                    case BOOLEAN:
-                        return operator.castTo(com.starrocks.catalog.Type.BOOLEAN);
-                    case DATE:
-                        return operator.castTo(com.starrocks.catalog.Type.DATE);
-                    case TIMESTAMP:
-                        return operator.castTo(com.starrocks.catalog.Type.DATETIME);
-                    case STRING:
-                    case UUID:
-                        // num and string has different comparator
-                        if (operator.getType().isNumericType()) {
-                            return null;
-                        }
-                        return operator.castTo(com.starrocks.catalog.Type.VARCHAR);
-                    case BINARY:
-                        return operator.castTo(com.starrocks.catalog.Type.VARBINARY);
+
+            Optional<ConstantOperator> res = Optional.empty();
+            switch (resultTypeID) {
+                case BOOLEAN:
+                    res = operator.castTo(com.starrocks.catalog.Type.BOOLEAN);
+                    break;
+                case DATE:
+                    res = operator.castTo(com.starrocks.catalog.Type.DATE);
+                    break;
+                case TIMESTAMP:
+                    res = operator.castTo(com.starrocks.catalog.Type.DATETIME);
+                    break;
+                case STRING:
+                case UUID:
+                    // num and string has different comparator
+                    if (operator.getType().isNumericType()) {
+                        return null;
+                    } else {
+                        res = operator.castTo(com.starrocks.catalog.Type.VARCHAR);
+                    }
+                    break;
+                case BINARY:
+                    res = operator.castTo(com.starrocks.catalog.Type.VARBINARY);
+                    break;
                     // num usually don't need cast, and num and string has different comparator
                     // cast is dangerous.
-                    case INTEGER:
-                    case LONG:
+                case INTEGER:
+                case LONG:
                     // usually not used as partition column, don't do much work
-                    case DECIMAL:
-                    case FLOAT:
-                    case DOUBLE:
-                    case STRUCT:
-                    case LIST:
-                    case MAP:
+                case DECIMAL:
+                case FLOAT:
+                case DOUBLE:
+                case STRUCT:
+                case LIST:
+                case MAP:
                     // not supported
-                    case FIXED:
-                    case TIME:
-                        return null;
-                }
-            } catch (Exception e) {
-                return null;
+                case FIXED:
+                case TIME:
+                    return null;
             }
-            return operator;
+
+            return res.isPresent() ? res.get() : null;
         }
 
         @Override