--- a/fe/fe-core/src/main/java/com/starrocks/server/MetadataMgr.java
+++ b/fe/fe-core/src/main/java/com/starrocks/server/MetadataMgr.java
@@ -24,12 +24,14 @@ import com.google.common.cache.RemovalCause;
 import com.google.common.cache.RemovalListener;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Lists;
 import com.starrocks.analysis.TableName;
 import com.starrocks.catalog.Column;
 import com.starrocks.catalog.Database;
 import com.starrocks.catalog.MaterializedIndexMeta;
 import com.starrocks.catalog.PartitionKey;
 import com.starrocks.catalog.Table;
+import com.starrocks.catalog.View;
 import com.starrocks.common.AlreadyExistsException;
 import com.starrocks.common.Config;
 import com.starrocks.common.DdlException;
@@ -47,6 +49,7 @@ import com.starrocks.connector.PartitionInfo;
 import com.starrocks.connector.RemoteFileInfo;
 import com.starrocks.connector.exception.StarRocksConnectorException;
 import com.starrocks.qe.ConnectContext;
+import com.starrocks.sql.analyzer.AnalyzerUtils;
 import com.starrocks.sql.ast.CreateTableStmt;
 import com.starrocks.sql.ast.DropTableStmt;
 import com.starrocks.sql.optimizer.OptimizerContext;
@@ -55,6 +58,7 @@ import com.starrocks.sql.optimizer.operator.scalar.ScalarOperator;
 import com.starrocks.sql.optimizer.statistics.ColumnStatistic;
 import com.starrocks.sql.optimizer.statistics.Statistics;
 import com.starrocks.thrift.TSinkCommitInfo;
+import org.apache.commons.collections4.CollectionUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jetbrains.annotations.NotNull;
@@ -274,12 +278,33 @@ public class MetadataMgr {
         connectorMetadata.ifPresent(metadata -> {
             try {
                 metadata.dropTable(stmt);
+                clearViewAnalyzedCache(ImmutableList.of(new TableName(dbName, tableName)));
             } catch (DdlException e) {
                 LOG.error("Failed to drop table {}.{}.{}", catalogName, dbName, tableName, e);
                 throw new StarRocksConnectorException("Failed to drop table %s.%s.%s. msg: %s",
                         catalogName, dbName, tableName, e.getMessage());
             }
         });
+    }
+
+    public static void clearViewAnalyzedCache(List<TableName> tableNames) {
+        List<Long> allDbIds = GlobalStateMgr.getCurrentState().getDbIds();
+        List<View> views = Lists.newArrayList();
+        for (Long viewDbId : allDbIds) {
+            Database db = GlobalStateMgr.getCurrentState().getDb(viewDbId);
+            if (null == db) {
+                continue;
+            }
+            db.getTables().stream().filter(View.class::isInstance).forEach(v -> views.add((View) v));
+        }
+
+        for (View view : views) {
+            Map<TableName, com.starrocks.catalog.Table> usedTable =
+                    AnalyzerUtils.collectAllTableAndView(view.getViewDefCache());
+            if (CollectionUtils.containsAny(usedTable.keySet(), tableNames)) {
+                view.clearAnalyzedCache();
+            }
+        }
     }
 
     public Optional<Table> getTable(TableName tableName) {