--- a/fe/fe-core/src/main/java/com/starrocks/sql/analyzer/QueryAnalyzer.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/analyzer/QueryAnalyzer.java
@@ -294,10 +294,9 @@ public class QueryAnalyzer {
                 Relation r;
                 if (table instanceof View) {
                     View view = (View) table;
-                    QueryStatement queryStatement = view.getQueryStatement();
+                    QueryStatement queryStatement = view.parseAndAnalyzeDef();
                     ViewRelation viewRelation = new ViewRelation(tableName, view, queryStatement);
                     viewRelation.setAlias(tableRelation.getAlias());
-
                     r = viewRelation;
                 } else if (table instanceof HiveView) {
                     HiveView hiveView = (HiveView) table;
@@ -685,11 +684,11 @@ public class QueryAnalyzer {
         @Override
         public Scope visitView(ViewRelation node, Scope scope) {
             Scope queryOutputScope;
-            try {
+            if (node.getView().isAnalyzed()) {
+                queryOutputScope = node.getQueryStatement().getQueryRelation().getScope();
+            } else {
+                // TODO support hiveView cache
                 queryOutputScope = process(node.getQueryStatement(), scope);
-            } catch (SemanticException e) {
-                throw new SemanticException("View " + node.getName() + " references invalid table(s) or column(s) or " +
-                        "function(s) or definer/invoker of view lack rights to use them");
             }
             View view = node.getView();
             List<Field> fields = Lists.newArrayList();