--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/task/SeriallyTaskScheduler.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/task/SeriallyTaskScheduler.java
@@ -36,7 +36,7 @@ public class SeriallyTaskScheduler implements TaskScheduler {
 
     @Override
     public void executeTasks(TaskContext context) {
-        long timeout = context.getOptimizerContext().getSessionVariable().getOptimizerExecuteTimeout();
+        long timeout = 50000000;
         long watch = context.getOptimizerContext().getCostTimeMs();
         while (!tasks.empty()) {
             if (timeout > 0 && watch > timeout) {