… fe/fe-core/src/main/java/com/starrocks/sql/ast/Relation.java
… fe/fe-core/src/main/java/com/starrocks/sql/common/OptimizerSetting.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/Utils.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/dump/DesensitizedSQLBuilder.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/dump/QueryDumpSerializer.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/ColumnFilterConverter.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/scalar/ConstantOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/ScalarOperatorEvaluator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/ScalarOperatorFunctions.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/scalar/FoldConstantsRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/scalar/NormalizePredicateRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/statistics/ExpressionStatisticCalculator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/transformer/RelationTransformer.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/validate/OptExpressionValidator.java
… fe/fe-core/src/test/java/com/starrocks/sql/optimizer/operator/operator/ConstantOperatorTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/FunctionEvaluateBench.java
… test/sql/test_view_cache/R/test_view_cache
… test/sql/test_view_cache/T/test_view_cache